# 🚀 FINAL FIX: Auto-Pilot Word Count Stuck & Model Selection